+++ project_management/agile/views.py
@@ -1,3 +1,20 @@
-from django.shortcuts import render
+# Created by Maximillian M. Estrada on 09-April-2020
+# Dockerfile to build Django Project Management
 
-# Create your views here.
+from rest_framework import generics
+from rest_framework.permissions import AllowAny
+
+from agile.serializers import *
+from agile.models import *
+
+
+class ValueListAPIView(generics.ListCreateAPIView):
+    queryset = Value.objects.all()
+    serializer_class = ValueSerializer
+    permission_classes = [AllowAny]
+
+
+class ValueDetailAPIView(generics.RetrieveUpdateAPIView):
+    queryset = Value.objects.all()
+    serializer_class = ValueSerializer
+    permission_classes = [AllowAny]