+++ project_management/project_management/urls.py
@@ -14,8 +14,14 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path
+from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
+    path('api/', include('rest_framework.urls')),
+
+    # Agile apps
+    path('api/agile/', include('agile.urls')),
 ]
+
+admin.site.site_title = admin.site.site_header = "Agile Project Management"